code/etc/born/requirements.txt
altgraph==0.16.1
arrow==0.10.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
binaryornot==0.4.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
constantly==15.1.0
cookiecutter==1.6.0
cryptography==2.1.4
deluge==1.3.15
dis3==0.1.3
enum34==1.1.6
future==0.15.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipaddress==1.0.17
Jinja2==2.10
jinja2-time==0.2.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.0
numpy==1.13.3
PAM==0.4.2
prompt-toolkit==1.0.14
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pygame===1.9.1release
Pygments==2.4.2
pygobject==3.26.1
PyInquirer==1.0.3
PyInstaller==3.5
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.6.1
python-hosts==0.4.7
python-libtorrent==1.1.5
pyxdg==0.25
PyYAML==3.12
regex==2019.8.19
requests==2.18.4
ruamel.yaml==0.15.34
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
Twisted==17.9.0
urllib3==1.22
wcwidth==0.1.7
whichcraft==0.4.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zope.interface==4.3.2
